[NFC][MemProf] Move IndexedMemProfData to its own header.
