+++ llvm/include/llvm/ProfileData/MemProf.h
@@ -842,57 +842,6 @@ struct LineLocation {
 
 // A pair of a call site location and its corresponding callee GUID.
 using CallEdgeTy = std::pair<LineLocation, uint64_t>;
-
-struct IndexedMemProfData {
-  // A map to hold memprof data per function. The lower 64 bits obtained from
-  // the md5 hash of the function name is used to index into the map.
-  llvm::MapVector<GlobalValue::GUID, IndexedMemProfRecord> Records;
-
-  // A map to hold frame id to frame mappings. The mappings are used to
-  // convert IndexedMemProfRecord to MemProfRecords with frame information
-  // inline.
-  llvm::MapVector<FrameId, Frame> Frames;
-
-  // A map to hold call stack id to call stacks.
-  llvm::MapVector<CallStackId, llvm::SmallVector<FrameId>> CallStacks;
-
-  FrameId addFrame(const Frame &F) {
-    const FrameId Id = hashFrame(F);
-    Frames.try_emplace(Id, F);
-    return Id;
-  }
-
-  CallStackId addCallStack(ArrayRef<FrameId> CS) {
-    CallStackId CSId = hashCallStack(CS);
-    CallStacks.try_emplace(CSId, CS);
-    return CSId;
-  }
-
-  CallStackId addCallStack(SmallVector<FrameId> &&CS) {
-    CallStackId CSId = hashCallStack(CS);
-    CallStacks.try_emplace(CSId, std::move(CS));
-    return CSId;
-  }
-
-private:
-  // Return a hash value based on the contents of the frame. Here we use a
-  // cryptographic hash function to minimize the chance of hash collisions.  We
-  // do persist FrameIds as part of memprof formats up to Version 2, inclusive.
-  // However, the deserializer never calls this function; it uses FrameIds
-  // merely as keys to look up Frames proper.
-  FrameId hashFrame(const Frame &F) const {
-    llvm::HashBuilder<llvm::TruncatedBLAKE3<8>, llvm::endianness::little>
-        HashBuilder;
-    HashBuilder.add(F.Function, F.LineOffset, F.Column, F.IsInlineFrame);
-    llvm::BLAKE3Result<8> Hash = HashBuilder.final();
-    FrameId Id;
-    std::memcpy(&Id, Hash.data(), sizeof(Hash));
-    return Id;
-  }
-
-  // Compute a CallStackId for a given call stack.
-  CallStackId hashCallStack(ArrayRef<FrameId> CS) const;
-};
 } // namespace memprof
 } // namespace llvm
