+++ llvm/lib/ProfileData/InstrProfWriter.cpp
@@ -18,7 +18,6 @@
 #include "llvm/IR/ProfileSummary.h"
 #include "llvm/ProfileData/IndexedMemProfData.h"
 #include "llvm/ProfileData/InstrProf.h"
-#include "llvm/ProfileData/MemProf.h"
 #include "llvm/ProfileData/ProfileCommon.h"
 #include "llvm/Support/Compression.h"
 #include "llvm/Support/Endian.h"
